requirements.txt
opencv-python
pandas
pytest
tqdm
matplotlib
imagehash
Pillow